+++ MagicAPI/src/main/java/com/elmakers/mine/bukkit/api/requirements/RequirementsProcessor.java
@@ -1,7 +1,9 @@
 package com.elmakers.mine.bukkit.api.requirements;
 
 import com.elmakers.mine.bukkit.api.magic.Mage;
-import org.bukkit.configuration.ConfigurationSection;
+
+import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
 
 /**
  * Implement this and register in LoadEvent to provide custom requirements.
@@ -14,18 +16,18 @@ public interface RequirementsProcessor {
      * There is no central definition of all known requirements, and processors need not register them in advance.
      * 
      * @param mage The mage being queried
-     * @param configuration The full configuration section provided in the requirement.
+     * @param requirement The requirement being checked
      * @return true if the player has the requirement
      */
-    boolean checkRequirement(Mage mage, ConfigurationSection configuration);
+    boolean checkRequirement(@Nonnull Mage mage, @Nonnull Requirement requirement);
 
     /**
      * This will be called when a requirement check has failed, in order to display to the player what they are missing.
      * This will generally be added to item lore for display in a GUI.
      * 
      * @param mage The mage being queried
-     * @param configuration The full configuration section provided in the requirement.
+     * @param requirement The requirement that failed
      * @return a player-readable description of what this requirement represents.
      */
-    String getRequirementDescription(Mage mage, ConfigurationSection configuration);
+    @Nullable String getRequirementDescription(@Nonnull Mage mage, @Nonnull Requirement requirement);
 }